pyproject.toml
[project]
name = "umip"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scipy>=1.11.0",
  "XlsxWriter>=3.1.0",
  "python-dateutil>=2.8.2",
  "pytrends>=4.9.2",
  "python-dotenv>=1.0.1"
]

[project.scripts]
umip = "umip.main:run"
